# src/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: AsapJoin Carpool Platform API
  description: |
    API REST complète pour la plateforme de covoiturage AsapJoin.
    Gère les trajets, réservations, livraisons de colis, paiements Stripe,
    wallet interne, politiques d'annulation, payouts et messagerie.
  version: 1.0.0
  contact:
    name: AsapJoin Team
servers:
  - url: http://localhost:3000
    description: Development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input
        traceId:
          type: string
          format: uuid
        details:
          type: array
          items:
            type: object
...
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        description:
          type: string
          nullable: true
        is_system:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserRole:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
            email:
              type: string
            first_name:
              type: string
            last_name:
              type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'

    # ─── Cancellation & Refund Schemas ───
...
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
paths:
...
  /api/v1/me/permissions:
    get:
      tags: [RBAC]
      summary: Get effective permissions for the current user
      responses:
        '200':
          description: Current permissions
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missing_permission:
                  value:
                    code: MISSING_PERMISSION
                    message: Missing required permission(s)
...
  /api/v1/admin/roles:
    get:
      tags: [RBAC]
      summary: List roles
      responses:
        '200':
          description: Role list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
    post:
      tags: [RBAC]
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code]
              properties:
                name:
                  type: string
                code:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Role created
        '403':
          description: Forbidden
        '409':
          description: Role code already exists
  /api/v1/admin/roles/{id}:
    get:
      tags: [RBAC]
      summary: Get role detail
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Role detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
    patch:
      tags: [RBAC]
      summary: Update a role
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Role updated
        '403':
          description: Role protected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                protected:
                  value:
                    code: ROLE_PROTECTED
                    message: System role is protected
    delete:
      tags: [RBAC]
      summary: Delete a role
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Role deleted
        '403':
          description: Role protected
  /api/v1/admin/roles/{id}/permissions:
    get:
      tags: [RBAC]
      summary: List permissions assigned to a role
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Role permissions
    put:
      tags: [RBAC]
      summary: Replace permissions assigned to a role
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissionIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Role permissions updated
        '403':
          description: Role protected
  /api/v1/admin/users/{id}/roles:
    get:
      tags: [RBAC]
      summary: List roles assigned to a user
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: User roles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRole'
    put:
      tags: [RBAC]
      summary: Replace roles assigned to a user
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: User roles updated
        '400':
          description: Invalid role assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_assignment:
                  value:
                    code: INVALID_ROLE_ASSIGNMENT
                    message: One or more roles do not exist

  # ─── Vehicles ───
...